requirements.txt
streamlit
speechrecognition
pytesseract
pillow
transformers
torch
